requirements.txt.txt
streamlit
opencv-python
numpy
Pillow
nltk
password-strength